requirements.txt
# Core audio processing
pyaudio>=0.2.11
sounddevice>=0.4.6
soundfile>=0.12.1
librosa>=0.10.1

# Speech recognition
openai-whisper>=20231117
faster-whisper>=0.10.0

# Text-to-speech
coqui-tts>=0.22.0

# AI integration
openai>=1.0.0

# Utilities
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0